+++ src/programs/index.ts
@@ -1,5 +1,4 @@
 import * as DailyChallenge from "./daily-challenge";
-import * as Game from "./game";
 import * as GroupManagerTools from "./group-manager";
 import GroupManager from "./group-manager";
 import * as MapTools from "./map";
@@ -9,7 +8,6 @@ import ExportManager from "./export-manager";
 
 export {
   ExportManager,
-  Game,
   GroupManager,
   GroupManagerTools,
   MapTools,
